up lesson-4

+++ lecture-demo/pallets/ocw-demo/src/lib.rs
@@ -480,7 +480,7 @@ impl<T: Trait> Module<T> {
 		let signer = Signer::<T, T::AuthorityId>::any_account();
 
 		// Translating the current block number to number and submit it on-chain
-		let number: u32 = block_number.try_into().unwrap_or(0);
+		let number: u32 = block_number.try_into().unwrap_or(0) as u32;
 
 		// `result` is in the type of `Option<(Account<T>, Result<(), ()>)>`. It is:
 		//   - `None`: no account is available for sending transaction
@@ -507,7 +507,7 @@ impl<T: Trait> Module<T> {
 	}
 
 	fn offchain_unsigned_tx(block_number: T::BlockNumber) -> Result<(), Error<T>> {
-		let number: u32 = block_number.try_into().unwrap_or(0);
+		let number: u32 = block_number.try_into().unwrap_or(0) as u32;
 		let call = Call::submit_number_unsigned(number);
 
 		// `submit_unsigned_transaction` returns a type of `Result<(), ()>`
@@ -523,7 +523,7 @@ impl<T: Trait> Module<T> {
 		// Retrieve the signer to sign the payload
 		let signer = Signer::<T, T::AuthorityId>::any_account();
 
-		let number: u32 = block_number.try_into().unwrap_or(0);
+		let number: u32 = block_number.try_into().unwrap_or(0) as u32;
 
 		// `send_unsigned_transaction` is returning a type of `Option<(Account<T>, Result<(), ()>)>`.
 		//   Similar to `send_signed_transaction`, they account for:
